refactor: vertical-writing: rename util
