+++ src/helpers/verticalWriting.js
